Bug #42480 ucslint: Re-implement 0014
to not use libtre, which implements the Levenshtein distance. Currently
we are using libtre with a custom built binding for Python. Debian
already carries a replacement call python[3]-levenshtein, but this is
overkill as we only check for the single word "univention" with all its
mis-spellings.
As there already are many exception, just roll our own implementation
using the plain Python 're' module.